+++ src/peeringdb/resource.py
@@ -3,9 +3,10 @@ PeeringDB resource definitions
 """
 
 from collections import OrderedDict
+from typing import Callable
 
 # Generate classes
-_NAMES = OrderedDict(
+_NAMES: OrderedDict[str, str] = OrderedDict(
     [
         ("org", "Organization"),
         ("campus", "Campus"),
@@ -23,12 +24,12 @@ _NAMES = OrderedDict(
     ]
 )
 
-RESOURCES_BY_TAG = OrderedDict()
+RESOURCES_BY_TAG: OrderedDict[str, type] = OrderedDict()
 
 for tag, name in _NAMES.items():
 
     class Meta(type):
-        def __repr__(cls, _name=name):
+        def __repr__(cls, _name: str = name) -> str:
             return _name
 
     Class = Meta(name, (), {"tag": tag})
@@ -36,15 +37,15 @@ for tag, name in _NAMES.items():
     RESOURCES_BY_TAG[tag] = Class
     globals()[name] = Class
 
-is_resource_tag = RESOURCES_BY_TAG.__contains__
+is_resource_tag: Callable[[str], bool] = RESOURCES_BY_TAG.__contains__
 
 
-def get_resource(tag):
+def get_resource(tag: str) -> type:
     return RESOURCES_BY_TAG[tag]
 
 
-get = get_resource
+get: Callable[[str], type] = get_resource
 
 
-def all_resources():
+def all_resources() -> list[type]:
     return list(RESOURCES_BY_TAG.values())